pyproject.toml
[project]
name = "cs787"
version = "1.0.0"
description = "CS787 Project"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "torch>=2.3.0",
  "torchvision>=0.20.0",
  "transformers>=4.44.0",
  "datasets>=2.14.0",
  "einops>=0.7.0",
  "diffusers>=0.30.0",
  "accelerate>=0.34.0",
  "huggingface-hub>=0.23.0",
  "safetensors>=0.4.2",
  "tokenizers>=0.19.1",
  "pandas>=2.1.0",
  "numpy>=1.26.0",
  "pillow>=10.3.0",
  "scikit-image>=0.22.0",
  "clean-fid>=0.1.35",
  "pyyaml>=6.0.1",
  "tqdm>=4.66.0",
  "wandb>=0.17.0",
  "pydicom>=2.4.4",
  "nibabel>=5.2.1",
  "scipy>=1.12.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.0",
]
